move partition writer to writer
